// src/hooks/services/seller/getSellersList/type.ts
export interface IParams {
  page?: string;
  sortBy?: string;
  sortOrder?: string;
  limit?: string;
  userId?: string;
  sellerName?: string;
  sellerType?: string;
  accountStatus?: string;
  kycStatus?: string;
  IsBankDetailsUpdated?: string;
  emailId?: string;
  mobileNumber?: string;
  accountId?: string;
  accountCreationDate?: string;
  lastUpdateTimestamp?: string;
}